MP0/requirements.txt
numpy
matplotlib
imageio
absl-py
